// yukka-ui/src/app/modules/profil/enums/TypPowiadomienia.ts
export enum TypPowiadomienia {
  KOMENTARZ_POST = 'KOMENTARZ_POST',
  WIADOMOSC_PRYWATNA = 'WIADOMOSC_PRYWATNA',
  POLUBIENIA_POST = 'POLUBIENIA_POST',
  OWOCOWANIE_ROSLIN = 'OWOCOWANIE_ROSLIN',
  OWOCOWANIE_ROSLIN_TERAZ = 'OWOCOWANIE_ROSLIN_TERAZ',
  KWITNIENIE_ROSLIN_TERAZ = 'KWITNIENIE_ROSLIN_TERAZ',
  PODLEWANIE_ROSLIN = 'PODLEWANIE_ROSLIN',
  USUNIECIE_KOMENTARZA = 'USUNIECIE_KOMENTARZA',
  BAN = 'BAN',
  ODBANOWANIE = 'ODBANOWANIE',
  OCENA_ROSLINY = 'OCENA_ROSLINY',
  USUNIECIE_ROSLINY = 'USUNIECIE_ROSLINY',
  GRATULACJE = 'GRATULACJE',
  SPECJALNE = 'SPECJALNE'
}